Struct-ify and completely rewrite the 2D tranforms
