+++ src/lib.rs
@@ -2,7 +2,6 @@ extern crate num_traits;
 extern crate itertools;
 
 pub mod dct_1d;
-pub mod dct_2d;
-
 pub mod haar_1d;
-pub mod haar_2d;
+
+pub mod transforms_2d;
